philo

+++ mandatory/ft_sleep.c
@@ -1,24 +1,23 @@
 #include "philo.h"
-#include <unistd.h>
 
-double	gettime()
+long	gettime()
 {
 	struct timeval	tv;
-	double			time;
+	long			time;
 
 	if (gettimeofday(&tv, NULL))
 	{
 		write(2, "gettimeofday failed\n", 21);
 		return (0);
 	}
-	time = (double)tv.tv_sec * 1000 + (double)tv.tv_usec / 1000;
+	time = tv.tv_sec * 1000 + tv.tv_usec / 1000;
 	return (time);
 }
 
 int	ft_sleep(long timetosleep)
 {
 	struct timeval	tv;
-	double	time;
+	long			time;
 
 	if (gettimeofday(&tv, NULL))
 	{
@@ -30,9 +29,3 @@ int	ft_sleep(long timetosleep)
 		usleep(100);
 	return (0);
 }
-
-// int	main()
-// {
-// 	int i = ft_sleep(10000);
-// 	return (0);
-// }
